requirements.txt
envdir
pandas
snowflake-connector-python
sqlalchemy
snowflake-connector-python[pandas]
snowflake-sqlalchemy
pysftp
retrying